JuliaBUGS.BUGSPrimitives.dnorm — Function

dnorm(μ, τ)

Return a Normal distribution object with mean μ and precision τ (the reciprocal of variance).

The mathematical form of the PDF for a Normal distribution in WinBUGS is given by:

\[f(x|μ,τ) = \sqrt{\frac{τ}{2π}} e^{-τ \frac{(x-μ)^2}{2}}\]

In Julia, this function uses mean $μ$ and standard deviation $σ = \sqrt{1 / τ}$.

JuliaBUGS.BUGSPrimitives.dlogis — Function

dlogis(μ, τ)

Return a Logistic distribution object with location parameter μ and scale parameter s, where $s = 1 / √τ$.

The mathematical form of the PDF for a Logistic distribution in WinBUGS is given by:

\[f(x|μ,τ) = \frac{\sqrt{τ} e^{-\sqrt{τ}(x-μ)}}{(1+e^{-\sqrt{τ}(x-μ)})^2}\]

JuliaBUGS.BUGSPrimitives.dt — Function

dt(μ, τ, ν)

Return a Student's t-distribution object with ν degrees of freedom, location μ, and scale $σ = √(1 / τ)$. If μ is 0 and σ is 1, the function returns a Tdist object. Otherwise, it returns a TDistShiftedScaled object.

The mathematical form of the PDF for a Student's t-distribution is given by:

\[f(x|ν,μ,σ) = \frac{Γ((ν+1)/2)}{Γ(ν/2) √{νπσ}} \left(1+\frac{1}{ν}\left(\frac{x-μ}{σ}\right)^2\right)^{-\frac{ν+1}{2}}\]

The mathematical form of the log-PDF for a Student's t-distribution is given by:

\[log(f(x|ν,μ,σ)) = log(Γ((ν+1)/2)) - log(Γ(ν/2)) - \frac{1}{2}log(νπσ) - \frac{ν+1}{2} log\left(1+\frac{1}{ν}\left(\frac{x-μ}{σ}\right)^2\right)\]

In the above, x is the random variable, ν is the degrees of freedom, μ is the location parameter, and σ is the scale parameter.

This implementation of the Student's t-distribution in Julia allows for a shift (determined by μ) and a scale (determined by σ) of the standard Student's t-distribution provided by the Distributions package.

Only pdf and logpdf are implemented for this distribution.

JuliaBUGS.BUGSPrimitives.ddexp — Function

ddexp(μ, τ)

Return a Laplace (Double Exponential) distribution object with location μ and scale $b = 1 / √τ$.

The mathematical form of the PDF for a Laplace distribution in WinBUGS is given by:

\[f(x|μ,τ) = \frac{\sqrt{τ}}{2} e^{-\sqrt{τ} |x-μ|}\]

JuliaBUGS.BUGSPrimitives.dflat — Function

dflat()

A distribution type representing a flat (uniform) prior over the real line. This is not a valid probability distribution, but can be used to represent a non-informative prior in Bayesian statistics. The cdf, logcdf, quantile, cquantile, rand, and rand methods are not implemented for this distribution, as they don't have meaningful definitions in the context of a flat prior. When use in a model, the parameters always need to be initialized to a valid value.

JuliaBUGS.BUGSPrimitives.dexp — Function

dexp(λ)

Return an Exponential distribution object with rate λ, where the rate is defined as $1 / λ$ in Julia's Distributions package.

The mathematical form of the PDF for an Exponential distribution in WinBUGS is given by:

\[f(x|λ) = λ e^{-λ x}\]

JuliaBUGS.BUGSPrimitives.dchisqr — Function

dchisqr(k)

Return a Chi-squared distribution object with k degrees of freedom.

The mathematical form of the PDF for a Chi-squared distribution in WinBUGS is given by:

\[f(x|k) = \frac{1}{2^{k/2} Γ(k/2)} x^{k/2 - 1} e^{-x/2}\]

JuliaBUGS.BUGSPrimitives.dweib — Function

dweib(a, b)

Return a Weibull distribution object with shape a and scale $λ = 1 / b$.

The mathematical form of the PDF for a Weibull distribution in WinBUGS is given by:

\[f(x|a,b) = \frac{b a (bx)^{a-1}}{e^{(bx)^a}}\]

JuliaBUGS.BUGSPrimitives.dlnorm — Function

dlnorm(μ, τ)

Return a LogNormal distribution object with location μ and scale $σ = 1 / √τ$.

The mathematical form of the PDF for a LogNormal distribution in WinBUGS is given by:

\[f(x|μ,τ) = \frac{\sqrt{τ}}{x\sqrt{2π}} e^{-τ/2 (\log(x) - μ)^2}\]

JuliaBUGS.BUGSPrimitives.dgamma — Function

dgamma(a, b)

Return a Gamma distribution object with shape a and scale $θ = 1 / b$.

The mathematical form of the PDF for a Gamma distribution in WinBUGS is given by:

\[f(x|a,b) = \frac{b^a}{Γ(a)} x^{a-1} e^{-bx}\]

JuliaBUGS.BUGSPrimitives.dpar — Function

dpar(a, b)

Return a Pareto distribution object with scale a and shape b.

The mathematical form of the PDF for a Pareto distribution in WinBUGS is given by:

\[f(x|a,b) = \frac{ba^b}{x^{b+1}}\]

JuliaBUGS.BUGSPrimitives.dgev — Function

dgev(μ, σ, η)

Return a GeneralizedExtremeValue distribution object with location μ, scale σ, and shape η.

The mathematical form of the PDF for a Generalized Extreme Value distribution in WinBUGS is given by:

\[f(x|μ,σ,η) = \frac{1}{σ} (1 + η ((x - μ)/σ))^{-1/η - 1} e^{-(1 + η ((x - μ)/σ))^{-1/η}}\]

JuliaBUGS.BUGSPrimitives.dgpar — Function

dgpar(μ, σ, η)

Return a GeneralizedPareto distribution object with location μ, scale σ, and shape η.

The mathematical form of the PDF for a Generalized Pareto distribution in WinBUGS is given by:

\[f(x|μ,σ,η) = \frac{1}{σ} (1 + η ((x - μ)/σ))^{-1/η - 1}\]

JuliaBUGS.BUGSPrimitives.df — Function

df(n::Real, m::Real, μ::Real=0, τ::Real=1)

Return an F-distribution object with n and m degrees of freedom, location μ, and scale τ. Raises a warning if μ ≠ 0 or τ ≠ 1, as these cases are not fully supported.

JuliaBUGS.BUGSPrimitives.dunif — Function

dunif(a, b)

Return a Uniform distribution object with lower bound a and upper bound b.

The mathematical form of the PDF for a Uniform distribution in WinBUGS is given by:

\[f(x|a,b) = \frac{1}{b - a}\]

JuliaBUGS.BUGSPrimitives.dbeta — Function

dbeta(a, b)

Return a Beta distribution object with shape parameters a and b.

The mathematical form of the PDF for a Beta distribution in WinBUGS is given by:

\[f(x|a,b) = \frac{x^{a-1} (1 - x)^{b-1}}{B(a, b)}\]

JuliaBUGS.BUGSPrimitives.dmnorm — Function

dmnorm(μ::Vector, T::Matrix)

Return a Multivariate Normal distribution object with mean vector μ and precision matrix T.

The mathematical form of the PDF for a Multivariate Normal distribution in WinBUGS is given by:

\[f(x|μ,T) = (2π)^{-k/2} |T|^{1/2} e^{-1/2 (x-μ)' T (x-μ)}\]

JuliaBUGS.BUGSPrimitives.dmt — Function

dmt(μ::Vector, T::Matrix, k)

Return a Multivariate T distribution object with mean vector μ, precision matrix T, and k degrees of freedom.

The mathematical form of the PDF for a Multivariate T distribution in WinBUGS is given by:

\[f(x|μ,T,k) = Γ((k+p)/2) / (Γ(k/2) (kπ)^{p/2} |T|^{1/2}) (1 + 1/k (x-μ)' T (x-μ))^{-((k+p)/2)}\]

JuliaBUGS.BUGSPrimitives.dwish — Function

dwish(R::Matrix, k)

Return a Wishart distribution object with k degrees of freedom and scale matrix R^(-1).

The mathematical form of the PDF for a Wishart distribution in WinBUGS is given by:

\[f(X|R,k) = |X|^{(k-p-1)/2} e^{-1/2 tr(RX)} / (2^{kp/2} |R|^{k/2} Γ_p(k/2))\]

JuliaBUGS.BUGSPrimitives.ddirich — Function

ddirich(θ::Vector)

Return a Dirichlet distribution object with parameters θ.

The mathematical form of the PDF for a Dirichlet distribution in WinBUGS is given by:

\[f(x|θ) = Γ(∑θ) / ∏Γ(θ) ∏x^{θ-1}\]

JuliaBUGS.BUGSPrimitives.dbern — Function

dbern(p)

Return a Bernoulli distribution object with success probability p.

The mathematical form of the PMF for a Bernoulli distribution in WinBUGS is given by:

\[P(x|p) = p^x (1 - p)^{1-x}\]

JuliaBUGS.BUGSPrimitives.dbin — Function

dbin(p, n)

Return a Binomial distribution object with number of trials n and success probability p.

The mathematical form of the PMF for a Binomial distribution in WinBUGS is given by:

\[P(x|n,p) = C(n, x) p^x (1 - p)^{n-x}\]

JuliaBUGS.BUGSPrimitives.dcat — Function

dcat(p)

Return a Categorical distribution object with probabilities p.

The mathematical form of the PMF for a Categorical distribution in WinBUGS is given by:

\[P(x|p) = p[x]\]

JuliaBUGS.BUGSPrimitives.dpois — Function

dpois(θ)

Return a Poisson distribution object with mean (and variance) θ.

The mathematical form of the PMF for a Poisson distribution in WinBUGS is given by:

\[P(x|θ) = e^{-θ} θ^x / x!\]

JuliaBUGS.BUGSPrimitives.dgeom — Function

dgeom(θ)

Return a Geometric distribution object with success probability θ.

The mathematical form of the PMF for a Geometric distribution in WinBUGS is given by:

\[P(x|θ) = (1 - θ)^{x-1} θ\]

JuliaBUGS.BUGSPrimitives.dnegbin — Function

dnegbin(p, r)

Return a Negative Binomial distribution object with number of failures r and success probability p.

The mathematical form of the PMF for a Negative Binomial distribution in WinBUGS is given by:

\[P(x|r,p) = C(x + r - 1, x) (1 - p)^x p^r\]

JuliaBUGS.BUGSPrimitives.dbetabin — Function

dbetabin(a, b, n)

Return a Beta Binomial distribution object with number of trials n and shape parameters a and b.

The mathematical form of the PMF for a Beta Binomial distribution in WinBUGS is given by:

\[P(x|a,b,n) = C(n, x) B(x + a, n - x + b) / B(a, b)\]

JuliaBUGS.BUGSPrimitives.dhyper — Function

dhyper(n1, n2, m1, ψ)

Return a Hypergeometric distribution object with total number of successes n1, total number of failures n2, and number of trials m1.

Only ψ = 1 is currently supported for hypergeometric distribution in this function.

The mathematical form of the PMF for a Hypergeometric distribution in WinBUGS is given by:

\[p(x | n1, n2, m1, \psi) = \frac{\binom{n1}{x} \binom{n2}{m1 - x} \psi^x}{\sum_{i=\max(0, m1-n2)}^{\min(n1,m1)} \binom{n1}{i} \binom{n2}{m1 - i} \psi^i}\]

JuliaBUGS.BUGSPrimitives.dmulti — Function

dmulti(θ::Vector, n)

Return a Multinomial distribution object with number of trials n and success probabilities θ.

The mathematical form of the PMF for a Multinomial distribution in WinBUGS is given by:

\[P(x|n,θ) = C(n, x) ∏θ^{x}\]