DynamicPPL.TestUtils

DynamicPPL.TestUtils.logjoint_true — Method

logjoint_true(model, θ)

Return the logjoint of model for θ.

Defaults to logprior_true(model, θ) + loglikelihood_true(model, θ).

This should generally be implemented by hand for every specific model so that the returned value can be used as a ground-truth for testing things like:

Validity of evaluation of model using a particular implementation of AbstractVarInfo.

Validity of a sampler when combined with DynamicPPL by running the sampler twice: once targeting ground-truth functions, e.g. logjoint_true, and once targeting model.

And more.

See also: logprior_true, loglikelihood_true.

DynamicPPL.TestUtils.loglikelihood_true — Function

loglikelihood_true(model, θ)

Return the loglikelihood of model for θ.

This should generally be implemented by hand for every specific model.

See also: logjoint_true, logprior_true.

DynamicPPL.TestUtils.logprior_true — Function

logprior_true(model, θ)

Return the logprior of model for θ.

This should generally be implemented by hand for every specific model.

See also: logjoint_true, loglikelihood_true.

DynamicPPL.TestUtils.test_sampler_continuous — Method

test_sampler_continuous([meanfunction, ]sampler, args...; kwargs...)

Test that sampler produces the correct marginal posterior means on all models in demo_models.

As of right now, this is just an alias for test_sampler_demo_models.

DynamicPPL.TestUtils.test_sampler_demo_models — Method

test_sampler_demo_models(meanfunction, sampler, args...; kwargs...)

Test that sampler produces the correct marginal posterior means on all models in demo_models.

In short, this method iterators through demo_models, calls AbstractMCMC.sample on the model and sampler to produce a chain, and then checks meanfunction(chain) against target provided in kwargs....

Arguments

meanfunction: A callable which computes the mean of the marginal means from the chain resulting from the sample call.

sampler: The AbstractMCMC.AbstractSampler to test.

args...: Arguments forwarded to sample.

Keyword arguments

target: Value to compare result of meanfunction(chain) to.

atol=1e-1: Absolute tolerance used in @test.

rtol=1e-3: Relative tolerance used in @test.

kwargs...: Keyword arguments forwarded to sample.