TuringBenchmarking.jl

A useful package for benchmarking and checking Turing.jl models.

Example

julia> using TuringBenchmarking, Turing
julia> @model function demo(x)
           s ~ InverseGamma(2, 3)
           m ~ Normal(0, sqrt(s))
           for i in 1:length(x)
               x[i] ~ Normal(m, sqrt(s))
           end
       enddemo (generic function with 2 methods)
julia> model = demo([1.5, 2.0]);
julia> benchmark_model(
           model;
           # Check correctness of computations
           check=true,
           # Automatic differentiation backends to check and benchmark
           adbackends=[:forwarddiff, :reversediff, :reversediff_compiled, :zygote]
       )┌ Warning: There is disagreement in the log-density values!
└ @ TuringBenchmarking ~/work/TuringBenchmarking.jl/TuringBenchmarking.jl/src/TuringBenchmarking.jl:248
┌──────────────────────────────────────┬─────────────┐
│                             Standard │ Log-density │
│                              backend │    distance │
├──────────────────────────────────────┼─────────────┤
│                ForwardDiff vs Zygote │        0.00 │
│ ForwardDiff vs ReverseDiff[compiled] │        0.59 │
│           ForwardDiff vs ReverseDiff │        0.00 │
│      Zygote vs ReverseDiff[compiled] │        0.59 │
│                Zygote vs ReverseDiff │        0.00 │
│ ReverseDiff[compiled] vs ReverseDiff │        0.59 │
└──────────────────────────────────────┴─────────────┘
┌ Warning: There is disagreement in the gradients!
└ @ TuringBenchmarking ~/work/TuringBenchmarking.jl/TuringBenchmarking.jl/src/TuringBenchmarking.jl:255
┌──────────────────────────────────────┬──────────┐
│                             Standard │ Gradient │
│                              backend │ distance │
├──────────────────────────────────────┼──────────┤
│                ForwardDiff vs Zygote │     0.00 │
│ ForwardDiff vs ReverseDiff[compiled] │     1.20 │
│           ForwardDiff vs ReverseDiff │     0.00 │
│      Zygote vs ReverseDiff[compiled] │     1.20 │
│                Zygote vs ReverseDiff │     0.00 │
│ ReverseDiff[compiled] vs ReverseDiff │     1.20 │
└──────────────────────────────────────┴──────────┘
2-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "evaluation" => 2-element BenchmarkTools.BenchmarkGroup:
	  tags: []
	  "linked" => Trial(500.000 ns)
	  "standard" => Trial(400.000 ns)
  "gradient" => 4-element BenchmarkTools.BenchmarkGroup:
	  tags: []
	  "Turing.Essential.ReverseDiffAD{false}()" => 2-element BenchmarkTools.BenchmarkGroup:
		  tags: ["ReverseDiff"]
		  "linked" => Trial(11.800 μs)
		  "standard" => Trial(11.200 μs)
	  "Turing.Essential.ReverseDiffAD{true}()" => 2-element BenchmarkTools.BenchmarkGroup:
		  tags: ["ReverseDiff[compiled]"]
		  "linked" => Trial(1.900 μs)
		  "standard" => Trial(1.900 μs)
	  "Turing.Essential.ForwardDiffAD{0, true}()" => 2-element BenchmarkTools.BenchmarkGroup:
		  tags: ["ForwardDiff"]
		  "linked" => Trial(800.000 ns)
		  "standard" => Trial(700.000 ns)
	  "Turing.Essential.ZygoteAD()" => 2-element BenchmarkTools.BenchmarkGroup:
		  tags: ["Zygote"]
		  "linked" => Trial(778.310 μs)
		  "standard" => Trial(768.010 μs)

API

TuringBenchmarking.benchmark_model

TuringBenchmarking.extract_stan_data

TuringBenchmarking.make_stan_suite

TuringBenchmarking.make_turing_suite

TuringBenchmarking.stan_model_string

TuringBenchmarking.benchmark_model — Method

benchmark_model(model::Turing.Model; suite_kwargs..., kwargs...)

Create and run a benchmark suite for model.

The benchmarking suite will be created using make_turing_suite. See make_turing_suite for the available keyword arguments and more information.

Keyword arguments

suite_kwargs: Keyword arguments passed to make_turing_suite.

kwargs: Keyword arguments passed to BenchmarkTools.run.

TuringBenchmarking.extract_stan_data — Function

extract_stan_data(model::DynamicPPL.Model)

Return the data in model in a format consumable by the corresponding Stan model.

The Stan model requires the return data to be either

A JSON string representing a dictionary with the data.

A path to a data file ending in .json.

TuringBenchmarking.make_stan_suite — Function

make_stan_suite(model::Turing.Model; kwargs...)

Create default benchmark suite for the Stan model corresponding to model.

TuringBenchmarking.make_turing_suite — Method

make_turing_suite(model::Turing.Model; kwargs...)

Create default benchmark suite for model.

Keyword arguments

adbackends: a collection of adbackends to use, specified either as a

Turing.Essential.ADBackend or using a Symbol. Defaults to Turing.Essential.ADBackend[Turing.Essential.ForwardDiffAD{0, true}(), Turing.Essential.ReverseDiffAD{false}(), Turing.Essential.ReverseDiffAD{true}(), Turing.Essential.ZygoteAD()].

run_once=true: if true, the body of each benchmark will be run once to avoid compilation to be included in the timings (this may occur if compilation runs longer than the allowed time limit).

check=false: if true, the log-density evaluations and the gradients will be compared against each other to ensure that they are consistent. Note that this will force run_once=true.

error_on_failed_check=false: if true, an error will be thrown if the check fails rather than just printing a warning, as is done by default.

error_on_failed_backend=false: if true, an error will be thrown if the evaluation of the log-density or the gradient fails for any of the backends rather than just printing a warning, as is done by default.

varinfo: the VarInfo to use. Defaults to DynamicPPL.VarInfo(model).

sampler: the Sampler to use. Defaults to nothing (i.e. no sampler).

context: the Context to use. Defaults to DynamicPPL.DefaultContext().

atol: the absolute tolerance to use for comparisons.

rtol: the relative tolerance to use for comparisons.

Notes

A separate "parameter" instance (DynamicPPL.VarInfo) will be created for each test. Hence if you have a particularly large model, you might want to only pass one adbackend at the time.

TuringBenchmarking.stan_model_string — Function

stan_model_string(model::DynamicPPL.Model)

Return a string defining the Stan model corresponding to model.