TuringBenchmarking

Documentation for TuringBenchmarking.

TuringBenchmarking.extract_stan_data

TuringBenchmarking.make_turing_suite

TuringBenchmarking.stan_model_string

TuringBenchmarking.extract_stan_data — Function

extract_stan_data(model::DynamicPPL.Model)

Return a Dict which can be consumed by the corresponding Stan model.

TuringBenchmarking.make_turing_suite — Method

make_turing_suite(model; kwargs...)

Create default benchmark suite for model.

Keyword arguments

adbackends: a collection of adbackends to use. Defaults to Turing.Essential.ADBackend[Turing.Essential.ForwardDiffAD{40, true}(), Turing.Essential.ForwardDiffAD{100, true}(), Turing.Essential.ZygoteAD(), Turing.Essential.ReverseDiffAD{false}(), Turing.Essential.ReverseDiffAD{true}()].

run_once=true: if true, the body of each benchmark will be run once to avoid compilation to be included in the timings (this may occur if compilation runs longer than the allowed time limit).

save_grads=false: if true and run_once is true, the gradients from the initial execution will be saved and returned as the second return-value. This is useful if you want to check correctness of the gradients for different backends.

Notes

A separate "parameter" instance (DynamicPPL.VarInfo) will be created for each test. Hence if you have a particularly large model, you might want to only pass one adbackend at the time.

TuringBenchmarking.stan_model_string — Function

stan_model_string(model::DynamicPPL.Model)

Return a string defining the Stan model corresponding to model.