API

JuliaBUGS.@bugs — Macro

@bugs(expr)

Convert Julia code to an Expr that can be used as the AST of a BUGS program. Checks that only allowed syntax is used, and normalizes certain expressions.

Used expression heads: :~ for tilde calls, :ref for indexing, :(:) for ranges. These are converted from :call variants.

@bugs(prog::String, replace_period=true, no_enclosure=false)

Produce similar output as @bugs, but takes a string as input. This is useful for parsing original BUGS programs.

Arguments

prog::String: The BUGS program code as a string.

replace_period::Bool: If true, periods in the BUGS code will be replaced (default true).

no_enclosure::Bool: If true, the parser will not expect the program to be wrapped between model{ } (default false).

JuliaBUGS.compile — Function

compile(model_def[, data, initializations])

Compile a BUGS model into a log density problem.

Arguments

model_def::Expr: The BUGS model definition.

data::NamedTuple or AbstractDict: The data to be used in the model. If none is passed, the data will be assumed to be empty.

initializations::NamedTuple or AbstractDict: The initial values for the model parameters. If none is passed, the parameters will be assumed to be initialized to zero.

Returns

A BUGSModel object representing the compiled model.

JuliaBUGS.BUGSModel — Type

BUGSModel

The BUGSModel object is used for inference and represents the output of compilation. It fully implements the LogDensityProblems.jl interface.

Fields

param_length::Int: The length of the parameters vector, defining the number of parameters in the model, store a tuple of integers, where the first integer is the length of the parameters vector in the original space, and the second integer is the length of the parameters vector in the transformed space.

var_lengths::Dict{VarName,Tuple{Int,Int}}: A dictionary that maps the names of the variables in the model to the corresponding lengths of the variables in the original space and the transformed space.

varinfo::SimpleVarInfo: An instance of DynamicPPL.SimpleVarInfo, specifically a dictionary that maps both data and value of variables in the model to the corresponding values.

parameters::Vector{VarName}: A vector containing the names of the parameters in the model. These parameters are defined to be stochastic variables that are not observed.

g::BUGSGraph: An instance of BUGSGraph, representing the dependency graph of the model.

sorted_nodes::Vector{VarName}: A vector containing the names of all the variables in the model, sorted in topological order.

JuliaBUGS.ConcreteNodeInfo — Type

ConcreteNodeInfo

Defines the information stored in each node of the BUGS graph, encapsulating the essential characteristics and functions associated with a node within the BUGS model's dependency graph.

Fields

node_type::VariableTypes: Specifies whether the node is a stochastic or logical variable.

node_function_expr::Expr: The node function expression.

node_args::Vector{VarName}: A vector containing the names of the variables that are arguments to the node function.

JuliaBUGS.BUGSGraph — Type

BUGSGraph

The BUGSGraph object represents the graph structure for a BUGS model. It is a type alias for MetaGraphsNext.MetaGraph with node type specified to ConcreteNodeInfo.