API

JuliaBUGS.Parser.@bugs — Macro

@bugs(prog::String, replace_period=true, no_enclosure=false)

Produce similar output as @bugs, but takes a string as input. This is useful for parsing original BUGS programs.

Arguments

prog::String: The BUGS program code as a string.

replace_period::Bool: If true, periods in the BUGS code will be replaced (default true).

no_enclosure::Bool: If true, the parser will not expect the program to be wrapped between model{ } (default false).

JuliaBUGS.compile — Function

compile(model_def[, data, initializations])

Compile a BUGS model into a log density problem.

Arguments

model_def::Expr: The BUGS model definition.

data::NamedTuple or AbstractDict: The data to be used in the model. If none is passed, the data will be assumed to be empty.

initializations::NamedTuple or AbstractDict: The initial values for the model parameters. If none is passed, the parameters will be assumed to be initialized to zero.

is_transformed::Bool=true: If true, the model parameters during inference will be transformed to the unconstrained space.

Returns

A BUGSModel object representing the compiled model.

JuliaBUGS.BUGSModel — Type

BUGSModel

The BUGSModel object is used for inference and represents the output of compilation. It fully implements the LogDensityProblems.jl interface.

Fields

transformed::Bool: Indicates whether the model parameters are in the transformed space.

untransformed_param_length::Int: The length of the parameters vector in the original space.

transformed_param_length::Int: The length of the parameters vector in the transformed space.

untransformed_var_lengths::Dict{VarName,Int}: A dictionary mapping the names of the variables to their lengths in the original space.

transformed_var_lengths::Dict{VarName,Int}: A dictionary mapping the names of the variables to their lengths in the transformed space.

varinfo::SimpleVarInfo: An instance of DynamicPPL.SimpleVarInfo, which is a dictionary-like data structure that maps both data and values of variables in the model to the corresponding values.

parameters::Vector{VarName}: A vector containing the names of the parameters in the model, defined as stochastic variables that are not observed. This vector should be consistent with sorted_nodes.

sorted_nodes::Vector{VarName}: A vector containing the names of all the variables in the model, sorted in topological order. In the case of a conditioned model, sorted_nodes include all the variables in parameters and the variables in the Markov blanket of parameters.

g::BUGSGraph: An instance of BUGSGraph, representing the dependency graph of the model.

base_model::Union{BUGSModel,Nothing}: If not Nothing, the model is a conditioned model; otherwise, it's the model returned by compile.

JuliaBUGS.ConcreteNodeInfo — Type

ConcreteNodeInfo

Defines the information stored in each node of the BUGS graph, encapsulating the essential characteristics and functions associated with a node within the BUGS model's dependency graph.

Fields

node_type::VariableTypes: Specifies whether the node is a stochastic or logical variable.

node_function_expr::Expr: The node function expression.

node_args::Vector{VarName}: A vector containing the names of the variables that are arguments to the node function.

JuliaBUGS.BUGSGraph — Type

BUGSGraph

The BUGSGraph object represents the graph structure for a BUGS model. It is a type alias for MetaGraphsNext.MetaGraph with node type specified to ConcreteNodeInfo.