JuliaBUGS.BUGSPrimitives.dnorm — Function

dnorm(μ, τ)

Construct a Normal distribution object in Julia using mean μ and precision τ as parameters.

In many statistical contexts, including the BUGS family of software, the Normal distribution is parametrized using precision, which is defined as the reciprocal of the variance. In contrast, Julia's Distributions package parametrizes the Normal distribution using mean (μ) and standard deviation (σ). This function accepts the mean μ and precision τ as inputs, then calculates the standard deviation as σ = √(1 / τ), and returns a Normal distribution object with mean μ and standard deviation σ.

The probability density function (PDF) of the Normal distribution as defined in the BUGS family of software is:

\[p(x|μ,τ) = \sqrt{\frac{τ}{2π}} e^{-τ \frac{(x-μ)^2}{2}}\]

In this equation, x is the random variable, μ is the mean of the distribution, and τ is the precision.

JuliaBUGS.BUGSPrimitives.dlogis — Function

dlogis(μ, τ)

Return a Logistic distribution object with location parameter μ and scale parameter s, where $s = 1 / √τ$.

The mathematical form of the PDF for a Logistic distribution in the BUGS family of softwares is given by:

\[p(x|μ,τ) = \frac{\sqrt{τ} e^{-\sqrt{τ}(x-μ)}}{(1+e^{-\sqrt{τ}(x-μ)})^2}\]

JuliaBUGS.BUGSPrimitives.dt — Function

dt(μ, τ, ν)

Construct a Student's t-distribution object with ν degrees of freedom, location μ, and scale $σ = √(1 / τ)$.

If μ is 0 and σ is 1, the function returns a TDist object from Distributions.jl. Otherwise, it returns a TDistShiftedScaled object.

The mathematical form of the PDF for a Student's t-distribution is given by:

\[p(x|ν,μ,σ) = \frac{Γ((ν+1)/2)}{Γ(ν/2) √{νπσ}} \left(1+\frac{1}{ν}\left(\frac{x-μ}{σ}\right)^2\right)^{-\frac{ν+1}{2}}\]

end

The mathematical form of the log-PDF for a Student's t-distribution is given by:

\[log(p(x|ν,μ,σ)) = log(Γ((ν+1)/2)) - log(Γ(ν/2)) - \frac{1}{2}log(νπσ) - \frac{ν+1}{2} log\left(1+\frac{1}{ν}\left(\frac{x-μ}{σ}\right)^2\right)\]

JuliaBUGS.BUGSPrimitives.TDistShiftedScaled — Type

TDistShiftedScaled(ν, μ, σ)

A Student's t-distribution object with ν degrees of freedom, location μ, and scale σ.

This struct allows for a shift (determined by μ) and a scale (determined by σ) of the standard Student's t-distribution provided by the Distributions package.

Only pdf and logpdf are implemented for this distribution.

See Also

TDist

JuliaBUGS.BUGSPrimitives.ddexp — Function

ddexp(μ, τ)

Return a Laplace (Double Exponential) distribution object with location μ and scale $b = 1 / √τ$.

The mathematical form of the PDF for a Laplace distribution in the BUGS family of softwares is given by:

\[p(x|μ,τ) = \frac{\sqrt{τ}}{2} e^{-\sqrt{τ} |x-μ|}\]

JuliaBUGS.BUGSPrimitives.dflat — Function

dflat()

A distribution type representing a flat (uniform) prior over the real line. This is not a valid probability distribution, but can be used to represent a non-informative prior in Bayesian statistics. The cdf, logcdf, quantile, cquantile, rand, and rand methods are not implemented for this distribution, as they don't have meaningful definitions in the context of a flat prior. When use in a model, the parameters always need to be initialized to a valid value.

JuliaBUGS.BUGSPrimitives.Flat — Type

Flat

Implement the flat distribution mimicking the behavior of the dflat distribution in the BUGS family of softwares.

JuliaBUGS.BUGSPrimitives.TruncatedFlat — Type

TruncatedFlat

Truncated version of the flat distribution.

JuliaBUGS.BUGSPrimitives.dexp — Function

dexp(λ)

Return an Exponential distribution object with rate λ, where the rate is defined as $1 / λ$ in Julia's Distributions package.

The mathematical form of the PDF for an Exponential distribution in the BUGS family of softwares is given by:

\[p(x|λ) = λ e^{-λ x}\]

JuliaBUGS.BUGSPrimitives.dchisqr — Function

dchisqr(k)

Return a Chi-squared distribution object with k degrees of freedom.

The mathematical form of the PDF for a Chi-squared distribution in the BUGS family of softwares is given by:

\[p(x|k) = \frac{1}{2^{k/2} Γ(k/2)} x^{k/2 - 1} e^{-x/2}\]

JuliaBUGS.BUGSPrimitives.dweib — Function

dweib(a, b)

Return a Weibull distribution object with shape parameter a and scale parameter $λ = 1 / b$.

The Weibull distribution is a common model for event times. The hazard or instantaneous risk of the event is (h(x) = abx^{a-1}). For a < 1 the hazard decreases with x; for a > 1 it increases. a = 1 results in the exponential distribution with constant hazard.

The mathematical form of the probability density function (PDF) for a Weibull distribution in the BUGS family of softwares is given by:

\[p(x|a,b) = abx^{a-1}e^{-b x^a}\]

JuliaBUGS.BUGSPrimitives.dlnorm — Function

dlnorm(μ, τ)

Return a LogNormal distribution object with location μ and scale $σ = 1 / √τ$.

The mathematical form of the PDF for a LogNormal distribution in the BUGS family of softwares is given by:

\[p(x|μ,τ) = \frac{\sqrt{τ}}{x\sqrt{2π}} e^{-τ/2 (\log(x) - μ)^2}\]

JuliaBUGS.BUGSPrimitives.dgamma — Function

dgamma(a, b)

Return a Gamma distribution object with shape a and scale $θ = 1 / b$.

The mathematical form of the PDF for a Gamma distribution in the BUGS family of softwares is given by:

\[p(x|a,b) = \frac{b^a}{Γ(a)} x^{a-1} e^{-bx}\]

JuliaBUGS.BUGSPrimitives.dpar — Function

dpar(a, b)

Return a Pareto distribution object with scale parameter b and shape parameter a.

The Pareto distribution, also known as the "80-20 rule", states that for many events, roughly 80% of the effects come from 20% of the causes. In terms of wealth distribution, it's often observed that 20% of the population owns 80% of a society's wealth.

The mathematical form of the probability density function (PDF) for a Pareto distribution in the BUGS family of softwares is given by:

\[p(x|a,b) = \frac{a b^a}{x^{a+1}}\]

In Julia, this function uses scale parameter b and shape parameter a to construct a Pareto(b, a) distribution object from the Distributions package.

JuliaBUGS.BUGSPrimitives.dgev — Function

dgev(μ, σ, η)

Return a GeneralizedExtremeValue distribution object with location μ, scale σ, and shape η.

The mathematical form of the PDF for a Generalized Extreme Value distribution in the BUGS family of softwares is given by:

\[p(x|μ,σ,η) = \frac{1}{σ} \left(1 + η \frac{x - μ}{σ}\right)^{-\frac{1}{η} - 1} e^{-\left(1 + η \frac{x - μ}{σ}\right)^{-\frac{1}{η}}}\]

where x is the random variable, μ is the location parameter, σ is the scale parameter, and η is the shape parameter. Note that the expression 1 + η ((x - μ)/σ) must be greater than zero for the function to be defined.

In Julia, this function returns a GeneralizedExtremeValue(μ, σ, η) distribution object from the Distributions package.

JuliaBUGS.BUGSPrimitives.dgpar — Function

dgpar(μ, σ, η)

Return a GeneralizedPareto distribution object with location μ, scale σ, and shape η.

The mathematical form of the PDF for a Generalized Pareto distribution in the BUGS family of softwares is given by:

\[p(x|μ,σ,η) = \frac{1}{σ} (1 + η ((x - μ)/σ))^{-1/η - 1}\]

JuliaBUGS.BUGSPrimitives.df — Function

df(n::Real, m::Real, μ::Real=0, τ::Real=1)

Return an F-distribution object with n and m degrees of freedom, location μ, and scale τ. Raises a warning if μ ≠ 0 or τ ≠ 1, as these cases are not fully supported.

The mathematical form of the PDF for an F-distribution in the BUGS family of softwares is given by:

\[p(x|n, m, μ, τ) = \frac{\Gamma\left(\frac{n+m}{2}\right)}{\Gamma\left(\frac{n}{2}\right) \Gamma\left(\frac{m}{2}\right)} \left(\frac{n}{m}\right)^{\frac{n}{2}} \sqrt{τ} \left(\sqrt{τ}(x - μ)\right)^{\frac{n}{2}-1} \left(1 + \frac{n \sqrt{τ}(x-μ)}{m}\right)^{-\frac{n+m}{2}}\]

where x is the random variable, n and m are the degrees of freedom, μ is the location parameter, and τ is the scale parameter. Note that the expression 1 + n sqrt(τ)(x - μ) / m must be greater than zero for the function to be defined.

In Julia, this function returns an FDist(n, m) distribution object from the Distributions package if μ = 0 and τ = 1.

JuliaBUGS.BUGSPrimitives.dunif — Function

dunif(a, b)

Return a Uniform distribution object with lower bound a and upper bound b.

The mathematical form of the PDF for a Uniform distribution in the BUGS family of softwares is given by:

\[p(x|a,b) = \frac{1}{b - a}\]

JuliaBUGS.BUGSPrimitives.dbeta — Function

dbeta(a, b)

Return a Beta distribution object with shape parameters a and b.

The mathematical form of the PDF for a Beta distribution in the BUGS family of softwares is given by:

\[p(x|a,b) = \frac{\Gamma(a + b)}{\Gamma(a)\Gamma(b)} x^{a-1} (1 - x)^{b-1}\]

JuliaBUGS.BUGSPrimitives.dmnorm — Function

dmnorm(μ::Vector, T::Matrix)

Return a Multivariate Normal distribution object with mean vector μ and precision matrix T.

The mathematical form of the PDF for a Multivariate Normal distribution in the BUGS family of softwares is given by:

\[p(x|μ,T) = (2π)^{-k/2} |T|^{1/2} e^{-1/2 (x-μ)' T (x-μ)}\]

JuliaBUGS.BUGSPrimitives.dmt — Function

dmt(μ::Vector, T::Matrix, k)

Return a Multivariate T distribution object with mean vector μ, precision matrix T, and k degrees of freedom.

The mathematical form of the PDF for a Multivariate T distribution in the BUGS family of softwares is given by:

\[p(x|k,μ,Σ) = \frac{\Gamma((k+p)/2)}{\Gamma(k/2) (k\pi)^{p/2} |Σ|^{1/2}} \left(1 + \frac{1}{k} (x-μ)^T Σ^{-1} (x-μ)\right)^{-\frac{k+p}{2}}\]

where x is the random variable, k is the degrees of freedom, μ is the mean vector, Σ is the scale matrix, and p is the dimension of x.

JuliaBUGS.BUGSPrimitives.dwish — Function

dwish(R::Matrix, k)

Return a Wishart distribution object with k degrees of freedom and scale matrix R^(-1).

The mathematical form of the PDF for a Wishart distribution in the BUGS family of softwares is given by:

\[p(X|R,k) = |X|^{(k-p-1)/2} e^{-(1/2) tr(RX)} / (2^{kp/2} |R|^{k/2} Γ_p(k/2))\]

where p is the dimension of X, and p should be less or equal to k.

This is the definition as in The BUGS Book (Lunn, D. J., Jackson, C., Best, N., Thomas, A., & Spiegelhalter, D. (2013). The BUGS Book: A Practical Introduction to Bayesian Analysis. CRC Press.), which is different from OpenBUGS' definition of the pdf of the Wishart distribution:

\[|R|^{k/2} |x|^{(k-p-1)/2} \exp\left(-\frac{1}{2} \text{Tr}(Rx)\right)\]

JuliaBUGS.BUGSPrimitives.ddirich — Function

ddirich(θ::Vector)

Return a Dirichlet distribution object with parameters θ.

The mathematical form of the PDF for a Dirichlet distribution in the BUGS family of softwares is given by:

\[p(x|θ) = \frac{Γ(\sum θ)}{∏ Γ(θ)} ∏ x_i^{θ_i - 1}\]

where x is a vector of random variables, each element x_i of which is between 0 and 1, and the elements of x sum up to 1. θ is a vector of parameters, each θ_i of which is greater than 0.

JuliaBUGS.BUGSPrimitives.dbern — Function

dbern(p)

Return a Bernoulli distribution object with success probability p.

The mathematical form of the PMF for a Bernoulli distribution in the BUGS family of softwares is given by:

\[p(x|p) = p^x (1 - p)^{1-x}\]

JuliaBUGS.BUGSPrimitives.dbin — Function

dbin(p, n)

Return a Binomial distribution object with number of trials n and success probability p.

The mathematical form of the PMF for a Binomial distribution in the BUGS family of softwares is given by:

\[p(x|n,p) = \binom{n}{x} p^x (1 - p)^{n-x}\]

end

where x is a random variable that can take the values from 0 to n.

JuliaBUGS.BUGSPrimitives.dcat — Function

dcat(p)

Return a Categorical distribution object with probabilities p.

The mathematical form of the PMF for a Categorical distribution in the BUGS family of softwares is given by:

\[p(x|p) = p[x]\]

JuliaBUGS.BUGSPrimitives.dpois — Function

dpois(θ)

Return a Poisson distribution object with mean (and variance) θ.

The mathematical form of the PMF for a Poisson distribution in the BUGS family of softwares is given by:

\[p(x|θ) = e^{-θ} θ^x / x!\]

JuliaBUGS.BUGSPrimitives.dgeom — Function

dgeom(θ)

Return a Geometric distribution object with success probability θ.

The mathematical form of the PMF for a Geometric distribution in the BUGS family of softwares is given by:

\[p(x|θ) = (1 - θ)^{x-1} θ\]

JuliaBUGS.BUGSPrimitives.dnegbin — Function

dnegbin(p, r)

Return a Negative Binomial distribution object with number of failures r and success probability p.

The mathematical form of the PMF for a Negative Binomial distribution in the BUGS family of softwares is given by:

\[P(x|r,p) = \binom{x + r - 1}{x} (1 - p)^x p^r\]

where x is a random variable that can take non-negative integer values.

JuliaBUGS.BUGSPrimitives.dbetabin — Function

dbetabin(a, b, n)

Return a Beta Binomial distribution object with number of trials n and shape parameters a and b.

The mathematical form of the PMF for a Beta Binomial distribution in the BUGS family of softwares is given by:

\[P(x|a, b, n) = \binom{n}{x} \binom{a + b - 1}{a + x - 1} / \binom{a + b + n - 1}{n}\]

where x is the number of successful trials, n is the total number of trials, a and b are the shape parameters of the Beta distribution.

JuliaBUGS.BUGSPrimitives.dhyper — Function

dhyper(n1, n2, m1, ψ)

Return a Hypergeometric distribution object. This distribution is used when sampling without replacement from a population consisting of n1 successes and n2 failures, with m1 being the number of trials or the sample size.

ψ is a scaling parameter which is currently unsupported in this function. Only ψ = 1 is currently supported.

The mathematical form of the PMF for a Hypergeometric distribution in the BUGS family of softwares is given by:

\[p(x | n1, n2, m1, \psi) = \frac{\binom{n1}{x} \binom{n2}{m1 - x} \psi^x}{\sum_{i=\max(0, m1-n2)}^{\min(n1,m1)} \binom{n1}{i} \binom{n2}{m1 - i} \psi^i}\]

In this formula, x is the number of successes in the sample, n1 is the total number of successes in the population, n2 is the total number of failures in the population, m1 is the number of trials, and \psi is a scaling parameter.

The sum in the denominator is over i from u0 to u1, where

\[u_0 = \max(0, m1 - n2), \quad u_1 = \min(n1, m1), \quad \text{and} \quad u_0 \leq x \leq u_1\]

JuliaBUGS.BUGSPrimitives.dmulti — Function

dmulti(θ::Vector, n)

Return a Multinomial distribution object with number of trials n and success probabilities θ.

The mathematical form of the PMF for a Multinomial distribution in the BUGS family of softwares is given by:

\[P(x|n,θ) = \frac{n!}{∏_{r} x_{r}!} ∏_{r} θ_{r}^{x_{r}}\]

where x is a vector of length R representing the count of successes in each of R categories, n is the total number of trials, and θ is a vector of length R representing the probability of success in each of R categories. The symbol ∏ denotes the product over all categories.