JuliaBUGS.BUGSPrimitives.abs — Function

abs(x)

Absolute value of x.

JuliaBUGS.BUGSPrimitives.cloglog — Function

cloglog(x)

Complementary log-log function of x. Can be used as link function.

\[\log{(-\log{(1 - x)})}\]

JuliaBUGS.BUGSPrimitives.cexpexp — Function

cexpexp(x)

Complementary exponential of the complementary exponential of x.

\[1 - \exp{(-\exp{(x)})}\]

JuliaBUGS.BUGSPrimitives.equals — Function

equals(x, y)

Returns 1 if $x$ is equal to $y$, 0 otherwise.

JuliaBUGS.BUGSPrimitives.exp — Function

exp(x)

Exponential of $x$.

JuliaBUGS.BUGSPrimitives.inprod — Function

inprod(a, b)

Inner product of $a$ and $b$.

JuliaBUGS.BUGSPrimitives.inverse — Function

inverse(m::AbstractMatrix)

Inverse of matrix $\mathbf{m}$.

JuliaBUGS.BUGSPrimitives.log — Function

log(x)

Natural logarithm of $x$.

JuliaBUGS.BUGSPrimitives.logistic — Function

logistic(x)

Logistic function of $x$.

\[\frac{1}{1 + \exp{(-x)}}\]

JuliaBUGS.BUGSPrimitives.logdet — Function

logdet(::AbstractMatrix)

Logarithm of the determinant of matrix $\mathbf{v}$.

JuliaBUGS.BUGSPrimitives.logfact — Function

logfact(x)

Logarithm of the factorial of $x$.

JuliaBUGS.BUGSPrimitives.loggam — Function

loggam(x)

Logarithm of the gamma function of $x$.

JuliaBUGS.BUGSPrimitives.icloglog — Function

icloglog(x)

Inverse complementary log-log function of $x$. Alias for cexpexp(x).

JuliaBUGS.BUGSPrimitives.logit — Function

logit(x)

Logit function of $x$.

\[\log{(\frac{x}{1 - x})}\]

JuliaBUGS.BUGSPrimitives.mexp — Function

mexp(x::AbstractMatrix)

Matrix exponential of $\mathbf{x}$.

JuliaBUGS.BUGSPrimitives.max — Function

max(args...)

Return the maximum value of the input arguments.

Statistics.mean — Function

mean(v::AbstractVector)

Return the mean of the input vector $\mathbf{v}$.

JuliaBUGS.BUGSPrimitives.min — Function

min(args...)

Return the minimum value of the input arguments.

JuliaBUGS.BUGSPrimitives.phi — Function

phi(x)

Cumulative distribution function (CDF) of the standard normal distribution evaluated at $x$.

JuliaBUGS.BUGSPrimitives.pow — Function

pow(a, b)

Return $a$ raised to the power of $b$.

JuliaBUGS.BUGSPrimitives.sqrt — Function

sqrt(x)

Return the square root of $x$.

JuliaBUGS.BUGSPrimitives.rank — Function

rank(v::AbstractVector, i::Integer)

Return the rank of the $i$-th element of $\mathbf{v}$.

JuliaBUGS.BUGSPrimitives.ranked — Function

ranked(v::AbstractVector, i::Integer)

Return the $i$-th element of $\mathbf{v}$ sorted in ascending order.

JuliaBUGS.BUGSPrimitives.round — Function

round(x)

Round $x$ to the nearest Integereger.

JuliaBUGS.BUGSPrimitives.sd — Function

sd(v::AbstractVector)

Return the standard deviation of the input vector $\mathbf{v}$.

JuliaBUGS.BUGSPrimitives.softplus — Function

softplus(x)

Return the softplus function of x, defined as $\log(1 + \exp(x))$.

JuliaBUGS.BUGSPrimitives.sort — Function

sort(v::AbstractVector)

Return a sorted copy of the input vector v.

JuliaBUGS.BUGSPrimitives._step — Function

_step(x)

Return 1 if $x$ is greater than 0, and 0 otherwise.

JuliaBUGS.BUGSPrimitives.sum — Function

sum(args...)

Return the sum of the input arguments.

JuliaBUGS.BUGSPrimitives.trunc — Function

trunc(x)

Return the Integereger part of $x$.

JuliaBUGS.BUGSPrimitives.sin — Function

sin(x)

Return the sine of $x$.

JuliaBUGS.BUGSPrimitives.arcsin — Function

arcsin(x)

Return the arcsine of $x$.

JuliaBUGS.BUGSPrimitives.arcsinh — Function

arcsinh(x)

Return the inverse hyperbolic sine of $x$.

JuliaBUGS.BUGSPrimitives.cos — Function

cos(x)

Return the cosine of $x$.

JuliaBUGS.BUGSPrimitives.arccos — Function

arccos(x)

Return the arccosine of $x$.

JuliaBUGS.BUGSPrimitives.arccosh — Function

arccosh(x)

Return the inverse hyperbolic cosine of $x$.

JuliaBUGS.BUGSPrimitives.tan — Function

tan(x)

Return the tangent of $x$.

JuliaBUGS.BUGSPrimitives.arctan — Function

arctan(x)

Return the arctangent of $x$.

JuliaBUGS.BUGSPrimitives.arctanh — Function

arctanh(x)

Return the inverse hyperbolic tangent of $x$.